feat: Refactoring the entire gui.py to split into several files instead
