+++ trade_route_tab.py
@@ -1,0 +1,239 @@
+import logging
+import sys
+from PyQt5.QtWidgets import QWidget, QVBoxLayout, QLabel, QLineEdit, QComboBox, QPushButton, QTableWidget, QMessageBox, QTableWidgetItem
+from PyQt5.QtCore import Qt
+import asyncio
+from api import API
+from config_manager import ConfigManager
+
+class TradeRouteTab(QWidget):
+    def __init__(self, main_widget):
+        super().__init__()
+        self.main_widget = main_widget
+        self.config_manager = ConfigManager()
+        self.api = API(
+            self.config_manager.get_api_key(),
+            self.config_manager.get_secret_key(),
+            self.config_manager.get_is_production(),
+            self.config_manager.get_debug()
+        )
+        self.logger = logging.getLogger(__name__)
+        self.terminals = []
+        self.initUI()
+        asyncio.ensure_future(self.load_systems())
+
+    def initUI(self):
+        layout = QVBoxLayout()
+
+        self.max_scu_input = QLineEdit()
+        self.max_scu_input.setPlaceholderText("Enter Max SCU")
+        layout.addWidget(QLabel("Max SCU:"))
+        layout.addWidget(self.max_scu_input)
+
+        self.max_investment_input = QLineEdit()
+        self.max_investment_input.setPlaceholderText("Enter Max Investment (UEC)")
+        layout.addWidget(QLabel("Max Investment (UEC):"))
+        layout.addWidget(self.max_investment_input)
+
+        self.departure_system_combo = QComboBox()
+        self.departure_system_combo.currentIndexChanged.connect(lambda: asyncio.ensure_future(self.update_planets()))
+        layout.addWidget(QLabel("Departure System:"))
+        layout.addWidget(self.departure_system_combo)
+
+        self.departure_planet_combo = QComboBox()
+        self.departure_planet_combo.currentIndexChanged.connect(lambda: asyncio.ensure_future(self.update_terminals()))
+        layout.addWidget(QLabel("Departure Planet:"))
+        layout.addWidget(self.departure_planet_combo)
+
+        terminal_label = QLabel("Select Terminal:")
+        self.terminal_filter_input = QLineEdit()
+        self.terminal_filter_input.setPlaceholderText("Filter Terminals")
+        self.terminal_filter_input.textChanged.connect(self.filter_terminals)
+        self.departure_terminal_combo = QComboBox()
+        layout.addWidget(terminal_label)
+        layout.addWidget(self.terminal_filter_input)
+        layout.addWidget(self.departure_terminal_combo)
+
+        find_route_button = QPushButton("Find Trade Route")
+        find_route_button.clicked.connect(lambda: asyncio.ensure_future(self.find_trade_routes()))
+        layout.addWidget(find_route_button)
+
+        self.trade_route_table = QTableWidget()
+        layout.addWidget(self.trade_route_table)
+
+        self.setLayout(layout)
+
+    async def load_systems(self):
+        try:
+            systems = await self.api.fetch_data("/star_systems")
+            for system in systems.get("data", []):
+                if system.get("is_available") == 1:
+                    self.departure_system_combo.addItem(system["name"], system["id"])
+            logging.info("Systems loaded successfully.")
+        except Exception as e:
+            logging.error(f"Failed to load systems: {e}")
+            QMessageBox.critical(self, "Error", f"Failed to load systems: {e}")
+
+    async def update_planets(self):
+        self.departure_planet_combo.clear()
+        self.departure_terminal_combo.clear()
+        self.terminal_filter_input.clear()  # Clear the filter input here
+        self.terminals = []
+        system_id = self.departure_system_combo.currentData()
+        if not system_id:
+            return
+        try:
+            planets = await self.api.fetch_data("/planets", params={'id_star_system': system_id})
+            for planet in planets.get("data", []):
+                self.departure_planet_combo.addItem(planet["name"], planet["id"])
+            logging.info("Planets loaded successfully.")
+        except Exception as e:
+            logging.error(f"Failed to load planets: {e}")
+            QMessageBox.critical(self, "Error", f"Failed to load planets: {e}")
+
+    async def update_terminals(self):
+        self.departure_terminal_combo.clear()
+        self.terminal_filter_input.clear()  # Ensure the filter input is cleared when updating terminals
+        self.terminals = []
+        planet_id = self.departure_planet_combo.currentData()
+        if not planet_id:
+            return
+        try:
+            terminals = await self.api.fetch_data("/terminals", params={'id_planet': planet_id})
+            self.terminals = [terminal for terminal in terminals.get("data", []) if terminal.get("type") == "commodity" and terminal.get("is_available") == 1]
+            self.filter_terminals()
+            logging.info("Terminals loaded successfully.")
+        except Exception as e:
+            logging.error(f"Failed to load terminals: {e}")
+            QMessageBox.critical(self, "Error", f"Failed to load terminals: {e}")
+
+    def filter_terminals(self):
+        filter_text = self.terminal_filter_input.text().lower()
+        self.departure_terminal_combo.clear()
+        for terminal in self.terminals:
+            if filter_text in terminal["name"].lower():
+                self.departure_terminal_combo.addItem(terminal["name"], terminal["id"])
+
+    async def find_trade_routes(self):
+        self.logger.log(logging.INFO, "Searching for a new Trade Route")
+        self.trade_route_table.setRowCount(0)  # Clear previous results
+
+        # Define the columns
+        columns = [
+            "Destination", "Commodity", "Buy SCU", "Buy Price", "Sell Price",
+            "Investment", "Unit Margin", "Total Margin", "Departure SCU Available",
+            "Arrival Demand SCU", "Profit Margin", "Arrival Terminal MCS"
+        ]
+        self.trade_route_table.setColumnCount(len(columns))
+        self.trade_route_table.setHorizontalHeaderLabels(columns)
+
+        try:
+            max_scu = int(self.max_scu_input.text()) if self.max_scu_input.text() else sys.maxsize
+            max_investment = float(self.max_investment_input.text()) if self.max_investment_input.text() else sys.maxsize
+            departure_system_id = self.departure_system_combo.currentData()
+            departure_planet_id = self.departure_planet_combo.currentData()
+            departure_terminal_id = self.departure_terminal_combo.currentData()
+
+            # Basic input validation
+            if not all([departure_system_id, departure_planet_id, departure_terminal_id]):
+                QMessageBox.warning(self, "Input Error", "Please Select Departure System, Planet, and Terminal.")
+                return
+
+            trade_routes = []
+            departure_commodities = await self.api.fetch_data("/commodities_prices", params={'id_terminal': departure_terminal_id})
+            self.logger.log(logging.INFO, f"Iterating through {len(departure_commodities.get('data', []))} commodities at departure terminal")
+            for departure_commodity in departure_commodities.get("data", []):
+                # Only get arrival terminals for commodities that can be bought in departure
+                if departure_commodity.get("price_buy") == 0:
+                    continue
+
+                arrival_commodities = await self.api.fetch_data("/commodities_prices", params={'id_commodity': departure_commodity.get("id_commodity")})
+                self.logger.log(logging.INFO, f"Found {len(arrival_commodities.get('data', []))} terminals that might sell {departure_commodity.get('commodity_name')}")
+
+                for arrival_commodity in arrival_commodities.get("data", []):
+                    # Check if terminal is available
+                    if arrival_commodity.get("is_available") == 0:
+                        continue
+
+                    # Check if terminal is the same as departure
+                    if arrival_commodity.get("id_terminal") == departure_terminal_id:
+                        continue
+
+                    buy_price = departure_commodity.get("price_buy", 0)
+                    available_scu = departure_commodity.get("scu_buy", 0)
+
+                    # Calculate trade route details
+                    sell_price = arrival_commodity.get("price_sell", 0)
+                    demand_scu = arrival_commodity.get("scu_sell_stock", 0) - arrival_commodity.get("scu_sell_users", 0)
+
+                    # Skip if buy or sell price is 0 or if SCU requirements aren't met
+                    if not buy_price or not sell_price or available_scu <= 0 or not demand_scu:
+                        continue
+
+                    max_buyable_scu = min(max_scu, available_scu, int(max_investment // buy_price), demand_scu)
+                    if max_buyable_scu <= 0:
+                        continue
+
+                    investment = buy_price * max_buyable_scu
+                    unit_margin = (sell_price - buy_price)
+                    total_margin = unit_margin * max_buyable_scu
+                    profit_margin = unit_margin / buy_price
+
+                    # Fetch arrival terminal data
+                    arrival_terminal = await self.api.fetch_data("/terminals", params={'id': arrival_commodity.get("id_terminal")})
+                    arrival_terminal_mcs = arrival_terminal.get("data")[0].get("mcs")
+
+                    trade_routes.append({
+                        "destination": next(
+                            (system["name"] for system in (await self.api.fetch_data("/star_systems")).get("data", [])
+                             if system["id"] == arrival_commodity.get("id_star_system")),
+                            "Unknown System"
+                        ) + " - " + next(
+                            (planet["name"] for planet in (await self.api.fetch_data("/planets", params={'id_star_system': arrival_commodity.get("id_star_system")})).get("data", [])
+                             if planet["id"] == arrival_commodity.get("id_planet")),
+                            "Unknown Planet"
+                        ) + " / " + arrival_commodity.get("terminal_name"),
+                        "commodity": departure_commodity.get("commodity_name"),
+                        "buy_scu": str(max_buyable_scu) + " SCU",
+                        "buy_price": str(buy_price) + " UEC",
+                        "sell_price": str(sell_price) + " UEC",
+                        "investment": str(investment) + " UEC",
+                        "unit_margin": str(unit_margin) + " UEC",
+                        "total_margin": str(total_margin) + " UEC",
+                        "departure_scu_available": str(available_scu) + " SCU",
+                        "arrival_demand_scu": str(demand_scu) + " SCU",
+                        "profit_margin": str(round(profit_margin * 100)) + "%",
+                        "arrival_terminal_mcs": arrival_terminal_mcs
+                    })
+                    self.trade_route_table.insertRow(len(trade_routes) - 1)
+                    for j, value in enumerate(trade_routes[len(trade_routes) - 1].values()):
+                        i = len(trade_routes) - 1
+                        item = QTableWidgetItem(str(value))
+                        item.setFlags(item.flags() & ~Qt.ItemIsEditable)  # Make the item non-editable
+                        self.trade_route_table.setItem(i, j, item)
+
+            # Sort trade routes by profit margin (descending)
+            trade_routes.sort(key=lambda x: float(x["total_margin"].split()[0]), reverse=True)
+
+            # Display up to the top 10 results
+            self.trade_route_table.setRowCount(0)  # Clear the table before adding sorted results
+            for i, route in enumerate(trade_routes[:10]):
+                self.trade_route_table.insertRow(i)
+                for j, value in enumerate(route.values()):
+                    item = QTableWidgetItem(str(value))
+                    item.setFlags(item.flags() & ~Qt.ItemIsEditable)  # Make the item non-editable
+                    self.trade_route_table.setItem(i, j, item)
+
+            if len(trade_routes) == 0:
+                self.trade_route_table.insertRow(0)
+                item = QTableWidgetItem("No results found")
+                item.setFlags(item.flags() & ~Qt.ItemIsEditable)  # Make the item non-editable
+                self.trade_route_table.setItem(0, 0, item)
+
+            # Resize columns to fit contents
+            self.trade_route_table.resizeColumnsToContents()
+
+            self.logger.log(logging.INFO, "Finished calculating Trade routes")
+        except Exception as e:
+            self.logger.log(logging.ERROR, f"An error occurred while finding trade routes: {e}")
+            QMessageBox.critical(self, "Error", f"An error occurred: {e}")
